fix: remove callbacks
to fix the problem of not being able to take over the callback func when passing it over to worker as an Object.
fix: add EXECUTE_ALL_CHARACTER type to WorkerMessage interface
fix: getImage() has been changed to return the url path of the image, not the callback.
fix: Updated switch for event handling in codeExecutor.ts
fix: Added a code that converts the parameter to fit the type when the worker received it as Object